Fix conflicts related to parameter parsing
